requirements-dev.txt
appdirs==1.4.4
black==20.8b1
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
envparse==0.2.0
filelock==3.0.12
flake8==3.9.0
identify==2.2.3
isort==5.8.0
loguru==0.5.3
logutil==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.8.1
pip==21.0.1
pre-commit==2.12.0
psutil==5.8.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.0
PyYAML==5.4.1
regex==2021.4.4
schedule==1.1.0
setuptools==56.0.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
virtualenv==20.4.3
wheel==0.36.2
